Organize related code into classes in OOP style
