+++ src/main/java/grok/Grok.java
@@ -1,40 +1,31 @@
 package grok;
 
+import storage.Storage;
+import tasks.Deadline;
+import tasks.Event;
+import tasks.Task;
+import tasks.Todo;
+
 import java.util.ArrayList;
 import java.util.Scanner;
 
 public class Grok {
     private static final String TEXT_FILE_DIRECTORY = "./data/duke.txt";
 
-    /**
-     * Pad a provided message with horizontal lines and indentation to separate bot input from user input
-     * @param msg - Message the bot should return to the user
-     * @return Standard bot message formatted message, easily distinguishable from user messages
-     */
-    private static String padMessage(String msg) {
-        // first stage - add horizontal lines above and below the message
-        String horizontalLine = "_".repeat(80);
-        String msgWithHLines = String.join("\n", horizontalLine, msg, horizontalLine);
-
-        // second stage - provide 1 indentation to all lines of the message, including the horizontal line
-        String indentSpaces = " ".repeat(4);
-
-        return indentSpaces.concat(msgWithHLines.replace("\n", "\n".concat(indentSpaces)));
-    }
-
-    // Provide information on added task and total number of tasks for the user
-    private static String addTaskMessage(Task t, ArrayList<Task> tasks) {
-        return "Got it. I've added this task:\n  " + t + "\nNow you have " + tasks.size() + " tasks in the list.";
+    public Grok(String filePath) {
+        ArrayList<Task> tasks = Storage.parseTextFile(TEXT_FILE_DIRECTORY);
     }
 
     public static void main(String[] args) {
-        ArrayList<Task> tasks = TextFileReader.parseFile(TEXT_FILE_DIRECTORY);
+
+
 
         // If data changes, prompt a save to text file
         boolean hasUnsavedChanges = false;
 
         Scanner scanner = new Scanner(System.in);
         System.out.println(padMessage("Hello! I'm Grok\nWhat ya wanna do to grok your way to success?"));
+
         while (true) {
             String userInput = scanner.nextLine();
             if (userInput.isEmpty()) {
