+++ src/main/java/tasks/Deadline.java
@@ -1,4 +1,6 @@
-package grok;
+package tasks;
+
+import grok.GrokInvalidUserInputException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
