+++ src/main/java/tasks/Event.java
@@ -1,4 +1,6 @@
-package grok;
+package tasks;
+
+import grok.GrokInvalidUserInputException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
